move token work to its own file and test it
